update: ID to uint in DTOs

+++ dto/models/administrativo/Token.go
@@ -3,5 +3,5 @@ package administrativo
 type Token struct {
 	Token    string "redis:token"
 	Refresh  string "redis:refresh_id"
-	AccessID uint64 "redis:access_id"
+	AccessID uint   "redis:access_id"
 }
