+++ dto/models/administrativo/Usuario.go
@@ -21,8 +21,8 @@ const (
 
 type Usuario struct {
 	gorm.Model
-	PerfilID uint64
-	ID       uint64 `gorm:"unique;primaryKey;autoIncrement" json:"ID"`
+	PerfilID uint
+	ID       uint   `gorm:"unique;primaryKey;autoIncrement" json:"ID"`
 	Nome     string `gorm:"size:255;not null;unique" json:"nome"`
 	Email    string `gorm:"size:100;not null,email;" json:"email"`
 	Senha    string `gorm:"size:100;not null;" json:"-"`
@@ -30,7 +30,7 @@ type Usuario struct {
 	Perfil   Perfil `gorm:"foreignKey:PerfilID,references:ID" json:"perfil" validate:"required"`
 }
 
-func (u *Usuario) Create(db *gorm.DB) (uint64, error) {
+func (u *Usuario) Create(db *gorm.DB) (uint, error) {
 	if verr := u.Validate("insert"); verr != nil {
 		return 0, verr
 	}
@@ -42,7 +42,7 @@ func (u *Usuario) Create(db *gorm.DB) (uint64, error) {
 	return u.ID, nil
 }
 
-func (u *Usuario) Update(db *gorm.DB, ID uint64) (*Usuario, error) {
+func (u *Usuario) Update(db *gorm.DB, ID uint) (*Usuario, error) {
 
 	if verr := u.Validate("insert"); verr != nil {
 		return nil, verr
@@ -83,7 +83,7 @@ func (u *Usuario) List(db *gorm.DB) (*[]Usuario, error) {
 }
 
 /*
-func (u *Usuario) Find(db *gorm.DB, ID uint64) (*Usuario, error) {
+func (u *Usuario) Find(db *gorm.DB, ID uint) (*Usuario, error) {
 
 	err := db.Debug().Model(Usuario{}).Where("id = ?", ID).Take(&u).Error
 	if err != nil {
@@ -118,7 +118,7 @@ func (u *Usuario) Find(db *gorm.DB, param string, ID string) (*Usuario, error) {
 
 	return u, nil
 }
-func (u *Usuario) Delete(db *gorm.DB, ID uint64) (int64, error) {
+func (u *Usuario) Delete(db *gorm.DB, ID uint) (int64, error) {
 	db = db.Debug().Where("id = ?", ID).Delete(&Usuario{})
 	if db.Error != nil {
 		return 0, db.Error
