+++ handlers/PersistenceHandler.go
@@ -11,11 +11,11 @@ type Tables interface {
 }
 
 type PersistenceHandler[T Tables] interface {
-	Create(db *gorm.DB) (uint64, error)
+	Create(db *gorm.DB) (uint, error)
 	List(db *gorm.DB) (*[]T, error)
-	Update(db *gorm.DB, ID uint64) (*T, error)
+	Update(db *gorm.DB, ID uint) (*T, error)
 	Find(db *gorm.DB, param string, value string) (*T, error)
-	Delete(db *gorm.DB, ID uint64) (int64, error)
+	Delete(db *gorm.DB, ID uint) (int64, error)
 	//FindBy(db *gorm.DB, param string, ID interface{}) (*T, error)
-	//DeleteBy(db *gorm.DB, cond string, ID uint64) (int64, error)
+	//DeleteBy(db *gorm.DB, cond string, ID uint) (int64, error)
 }
